vec_warnings.cpp and warnings.cpp merge
