+++ sycl/include/CL/sycl/types.hpp
@@ -331,7 +331,7 @@ convertImpl(T Value) {
                           !std::is_same<OpenCLT, OpenCLR>::value &&            \
                           (std::is_same<OpenCLR, DestType>::value ||           \
                            (std::is_same<OpenCLR, signed char>::value &&       \
-                               std::is_same<DestType, char>::value)),          \
+                            std::is_same<DestType, char>::value)),             \
                       R>                                                       \
   convertImpl(T Value) {                                                       \
     OpenCLT OpValue = cl::sycl::detail::convertDataToType<T, OpenCLT>(Value);  \
@@ -383,7 +383,7 @@ convertImpl(T Value) {
   detail::enable_if_t<is_sint_to_float<T, R>::value &&                         \
                           (std::is_same<OpenCLR, DestType>::value ||           \
                            (std::is_same<OpenCLR, _Float16>::value &&          \
-                               std::is_same<DestType, half>::value)),          \
+                            std::is_same<DestType, half>::value)),             \
                       R>                                                       \
   convertImpl(T Value) {                                                       \
     OpenCLT OpValue = cl::sycl::detail::convertDataToType<T, OpenCLT>(Value);  \
@@ -403,7 +403,7 @@ __SYCL_GENERATE_CONVERT_IMPL(SToF, double)
   detail::enable_if_t<is_uint_to_float<T, R>::value &&                         \
                           (std::is_same<OpenCLR, DestType>::value ||           \
                            (std::is_same<OpenCLR, _Float16>::value &&          \
-                               std::is_same<DestType, half>::value)),          \
+                            std::is_same<DestType, half>::value)),             \
                       R>                                                       \
   convertImpl(T Value) {                                                       \
     OpenCLT OpValue = cl::sycl::detail::convertDataToType<T, OpenCLT>(Value);  \
@@ -424,8 +424,8 @@ __SYCL_GENERATE_CONVERT_IMPL(UToF, double)
   detail::enable_if_t<is_float_to_float<T, R>::value &&                        \
                           !std::is_same<OpenCLT, OpenCLR>::value &&            \
                           (std::is_same<OpenCLR, DestType>::value ||           \
-                          (std::is_same<OpenCLR, _Float16>::value &&           \
-                               std::is_same<DestType, half>::value)) &&        \
+                           (std::is_same<OpenCLR, _Float16>::value &&          \
+                            std::is_same<DestType, half>::value)) &&           \
                           RoundingModeCondition<roundingMode>::value,          \
                       R>                                                       \
   convertImpl(T Value) {                                                       \
